# hot-wheels-backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hot Wheels Collector API
  version: 1.0.0
  description: |
    # Hot Wheels Collector Platform API

    A comprehensive REST API for Hot Wheels collectors to manage their collections,
    trade with other collectors, and explore the complete Hot Wheels catalog.

    ## Features
    - **Authentication**: JWT-based secure authentication system
    - **Catalog**: Browse 50+ years of Hot Wheels models
    - **Collections**: Track and manage personal collections
    - **Marketplace**: Buy, sell, and trade with other collectors
    - **Reviews**: Rate and review transactions

    ## Getting Started
    1. Register a new account via `/api/v1/auth/register`
    2. Login to receive JWT tokens via `/api/v1/auth/login`
    3. Include the access token in the `Authorization` header
    4. Start exploring the catalog and managing your collection!

    ## Rate Limits
    - Anonymous users: 100 requests/minute
    - Authenticated users: 500 requests/minute
    - Rate limit headers included in all responses

    ## Support
    For API support, please contact the development team.

  contact:
    name: API Support
    email: support@hotwheels-collector.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.hotwheels-collector.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `/api/v1/auth/login`.
        Include in Authorization header: `Bearer <token>`
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: "Invalid request parameters"
            code:
              type: string
              example: "VALIDATION_ERROR"
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false

    # ==================== User Schemas ====================
...
    CarModel:
      type: object
      properties:
        model_id:
          type: integer
          example: 1
        casting_id:
          type: integer
          example: 1
        series_id:
          type: integer
          nullable: true
          example: 1
        release_year:
          type: integer
          example: 2024
        color:
          type: string
          example: "Spectraflame Red"
        tampo_design:
          type: string
          nullable: true
          example: "Racing stripes with #7"
        wheel_type:
          type: string
          nullable: true
          example: "5-Spoke"
        base_color:
          type: string
          nullable: true
          example: "Chrome"
        window_color:
          type: string
          nullable: true
          example: "Tinted"
        interior_color:
          type: string
          nullable: true
          example: "Black"
        production_code:
          type: string
          nullable: true
          example: "L2593"
        sku:
          type: string
          nullable: true
          example: "HCW35-N9B1"
        rarity_level:
          type: string
          enum: [Common, Uncommon, Rare, Chase, Super Treasure Hunt]
          example: "Common"
        estimated_production_quantity:
          type: integer
          nullable: true
          example: 50000
        msrp:
          type: number
          format: float
          nullable: true
          example: 1.29
        casting_name:
          type: string
          description: Included when relations are requested
          example: "'67 Camaro"
        series_name:
          type: string
          nullable: true
          description: Included when relations are requested
          example: "Mainline 2024"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CollectionItem:
      type: object
      properties:
        collection_id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        model_id:
          type: integer
          example: 1
        acquisition_date:
          type: string
          format: date
          nullable: true
          example: "2024-06-15"
        acquisition_price:
          type: number
          format: float
          nullable: true
          example: 4.99
        condition:
          type: string
          enum: [Mint, Near Mint, Excellent, Good, Fair, Poor]
          example: "Mint"
        is_in_package:
          type: boolean
          example: true
        package_condition:
          type: string
          enum: [Mint, Near Mint, Excellent, Good, Fair, Poor]
          nullable: true
          example: "Mint"
        quantity:
          type: integer
          minimum: 1
          example: 1
        storage_location:
          type: string
          nullable: true
          example: "Display Case A"
        notes:
          type: string
          nullable: true
          example: "Found at local Walmart"
        is_for_trade:
          type: boolean
          example: false
        is_for_sale:
          type: boolean
          example: false
        car_model:
          $ref: '#/components/schemas/CarModel'
          description: Included when relations are requested
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AddToCollectionRequest:
      type: object
      required:
        - model_id
      properties:
        model_id:
          type: integer
          example: 1
        acquisition_date:
          type: string
          format: date
          example: "2024-06-15"
        acquisition_price:
          type: number
          format: float
          example: 4.99
        condition:
          type: string
          enum: [Mint, Near Mint, Excellent, Good, Fair, Poor]
          default: "Mint"
          example: "Mint"
        is_in_package:
          type: boolean
          default: true
          example: true
        package_condition:
          type: string
          enum: [Mint, Near Mint, Excellent, Good, Fair, Poor]
          example: "Mint"
        quantity:
          type: integer
          minimum: 1
          default: 1
          example: 1
        storage_location:
          type: string
          example: "Display Case A"
        notes:
          type: string
          example: "Found at local Walmart"
        is_for_trade:
          type: boolean
          default: false
          example: false
        is_for_sale:
          type: boolean
          default: false
          example: false
    UpdateCollectionItemRequest:
      type: object
      properties:
        acquisition_price:
          type: number
          format: float
          example: 5.99
        condition:
          type: string
          enum: [Mint, Near Mint, Excellent, Good, Fair, Poor]
          example: "Near Mint"
        is_in_package:
          type: boolean
          example: false
        package_condition:
          type: string
          enum: [Mint, Near Mint, Excellent, Good, Fair, Poor]
          nullable: true
        quantity:
          type: integer
          minimum: 1
        storage_location:
          type: string
        notes:
          type: string
        is_for_trade:
          type: boolean
        is_for_sale:
          type: boolean
    WishlistItem:
      type: object
      properties:
        wishlist_id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        model_id:
          type: integer
          example: 1
        priority:
          type: integer
          minimum: 1
          maximum: 5
          example: 3
        max_price_willing:
          type: number
          format: float
          nullable: true
          example: 25.00
        notes:
          type: string
          nullable: true
          example: "Looking for loose or carded"
        car_model:
          $ref: '#/components/schemas/CarModel'
          description: Included when relations are requested
        created_at:
          type: string
          format: date-time
    AddToWishlistRequest:
      type: object
      required:
        - model_id
      properties:
        model_id:
          type: integer
          example: 1
        priority:
          type: integer
          minimum: 1
          maximum: 5
          example: 3
        max_price_willing:
          type: number
          format: float
          example: 25.00
        notes:
          type: string
          example: "Looking for loose or carded"
    CollectionStats:
      type: object
      properties:
        total_items:
          type: integer
          example: 150
        total_value:
          type: number
          format: float
          example: 750.50
        items_for_sale:
          type: integer
          example: 10
        items_for_trade:
          type: integer
          example: 25
        by_condition:
          type: object
          additionalProperties:
            type: integer
        by_rarity:
          type: object
          additionalProperties:
            type: integer

    # ==================== Marketplace Schemas ====================
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: "Missing required fields: email, password"
    Unauthorized:
      description: Unauthorized - authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: "Authentication required"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: "Resource not found"
...
paths:
...
  /api/v1/collections:
    get:
      summary: Get user's collection
      description: Returns the authenticated user's collection with optional filters
      tags:
        - Collections
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: for_sale
          in: query
          description: Filter items marked for sale
          schema:
            type: boolean
        - name: for_trade
          in: query
          description: Filter items marked for trade
          schema:
            type: boolean
        - name: condition
          in: query
          description: Filter by condition
          schema:
            type: string
            enum: [Mint, Near Mint, Excellent, Good, Fair, Poor]
      responses:
        '200':
          description: Collection items
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/CollectionItem'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Add to collection
      description: Adds a car model to the user's collection
      tags:
        - Collections
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToCollectionRequest'
            examples:
              minimal:
                summary: Minimal - just model ID
                value:
                  model_id: 1
              full:
                summary: Full details
                value:
                  model_id: 1
                  acquisition_date: "2024-06-15"
                  acquisition_price: 4.99
                  condition: "Mint"
                  is_in_package: true
                  package_condition: "Mint"
                  quantity: 1
                  storage_location: "Display Case A"
                  notes: "Found at local Walmart"
                  is_for_trade: false
                  is_for_sale: false
      responses:
        '201':
          description: Added to collection
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      collection_item:
                        $ref: '#/components/schemas/CollectionItem'
                  message:
                    type: string
                    example: "Added to collection successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/collections/{collection_id}:
    put:
      summary: Update collection item
      description: Updates a specific item in the user's collection
      tags:
        - Collections
      security:
        - BearerAuth: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCollectionItemRequest'
      responses:
        '200':
          description: Item updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      collection_item:
                        $ref: '#/components/schemas/CollectionItem'
                  message:
                    type: string
                    example: "Collection item updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Remove from collection
      description: Removes an item from the user's collection
      tags:
        - Collections
      security:
        - BearerAuth: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Item removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Item removed from collection"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/collections/stats:
    get:
      summary: Get collection statistics
      description: Returns statistics about the user's collection
      tags:
        - Collections
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Collection statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      stats:
                        $ref: '#/components/schemas/CollectionStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/collections/wishlist:
    get:
      summary: Get wishlist
      description: Returns the user's wishlist
      tags:
        - Collections
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Wishlist items
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/WishlistItem'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Add to wishlist
      description: Adds a car model to the user's wishlist
      tags:
        - Collections
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToWishlistRequest'
      responses:
        '201':
          description: Added to wishlist
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      wishlist_item:
                        $ref: '#/components/schemas/WishlistItem'
                  message:
                    type: string
                    example: "Added to wishlist successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/collections/wishlist/{wishlist_id}:
    delete:
      summary: Remove from wishlist
      description: Removes an item from the user's wishlist
      tags:
        - Collections
      security:
        - BearerAuth: []
      parameters:
        - name: wishlist_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Item removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Item removed from wishlist"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Marketplace ====================
...